pyproject.toml
[project]
name = "bastion"
version = "0.1.0"
description = "CLI tool for Linux server administration"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
bastion = "bastion.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bastion"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
